fix: minor re-org and output formatting
